tests/wpt/web-platform-tests/websockets/interfaces/WebSocket/constants/005.html

<!doctype html>
<title>WebSockets: defineProperty getter for constants</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.js?pipe=sub></script>
<meta name="variant" content="">
<meta name="variant" content="?wss">
<div id=log></div>
<script>
var constants = ['CONNECTING', 'OPEN', 'CLOSING', 'CLOSED'];
for (var i = 0; i < constants.length; ++i) {
  test(function() {
    assert_throws(new TypeError(), function() {
      Object.defineProperty(WebSocket.prototype, constants[i], {
        get: function() { return 'foo'; }
      });
    });
  }, "defineProperty getter " + constants[i]);
};
</script>
